+++ CHANGELOG.md
@@ -22,6 +22,8 @@
     - Changed "Add label" hotkey to Ctrl+G (#305)
 - Module [MIDI-KEY](./docs/MidiKey.md)
     - Added options for slot-specific key modifiers (#344)
+- Module [ORBIT](./docs/Orbit.md)
+    - Added output level control (#286)
 - Module [STRIP](./docs/Strip.md)
     - Remember last used folder for strips and selections on dialogs (#307)
 - Module [STROKE](./docs/Stroke.md)